requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
pymupdf>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
openai>=1.0.0
